moving source file to /src, removing old files
